pyproject.toml
[project]
name = "tratativa-base-455"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "selenium>=4.39.0",
    "webdriver-manager>=4.0.2",
]
